core: Split Reaction Ensemble methods
