+++ src/core/CMakeLists.txt
@@ -26,7 +26,6 @@ set(EspressoCore_SRC
     polymer.cpp
     pressure.cpp
     rattle.cpp
-    reaction_ensemble.cpp
     rotate_system.cpp
     rotation.cpp
     Observable_stat.cpp
@@ -102,6 +101,7 @@ add_subdirectory(io)
 add_subdirectory(nonbonded_interactions)
 add_subdirectory(object-in-fluid)
 add_subdirectory(observables)
+add_subdirectory(reaction_methods)
 add_subdirectory(virtual_sites)
 
 if(WITH_TESTS)
