+++ src/core/reaction_methods/WidomInsertion.hpp
@@ -1,0 +1,36 @@
+/*
+ * Copyright (C) 2010-2019 The ESPResSo project
+ *
+ * This file is part of ESPResSo.
+ *
+ * ESPResSo is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * ESPResSo is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+#ifndef REACTION_ENSEMBLE_WIDOM_INSERTION_HPP
+#define REACTION_ENSEMBLE_WIDOM_INSERTION_HPP
+
+#include "reaction_methods/ReactionAlgorithm.hpp"
+
+#include <utility>
+
+namespace ReactionMethods {
+
+/** Widom insertion method */
+class WidomInsertion : public ReactionAlgorithm {
+public:
+  WidomInsertion(int seed) : ReactionAlgorithm(seed) {}
+  std::pair<double, double> measure_excess_chemical_potential(int reaction_id);
+};
+
+} // namespace ReactionMethods
+#endif
