+++ src/core/reaction_methods/tests/reaction_methods_utils_test.cpp
@@ -17,13 +17,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-/* Unit tests for the ReactionEnsemble utility functions. */
+/* Unit tests for the ReactionMethods utility functions. */
 
-#define BOOST_TEST_MODULE ReactionEnsemble utility functions test
+#define BOOST_TEST_MODULE ReactionMethods utility functions test
 #define BOOST_TEST_DYN_LINK
 #include <boost/test/unit_test.hpp>
 
-#include "reaction_ensemble.hpp"
+#include "reaction_methods/utils.hpp"
 
 #include <cmath>
 #include <limits>
@@ -31,7 +31,7 @@
 #include <vector>
 
 BOOST_AUTO_TEST_CASE(find_minimum_non_negative_value_test) {
-  using namespace ReactionEnsemble;
+  using namespace ReactionMethods;
 
   BOOST_CHECK_EQUAL(find_minimum_non_negative_value({1, 2, 3}), 1);
   BOOST_CHECK_EQUAL(find_minimum_non_negative_value({3, 2, 1}), 1);
@@ -41,7 +41,7 @@ BOOST_AUTO_TEST_CASE(find_minimum_non_negative_value_test) {
 }
 
 BOOST_AUTO_TEST_CASE(average_list_of_allowed_entries_test) {
-  using namespace ReactionEnsemble;
+  using namespace ReactionMethods;
   constexpr double tol = 100 * std::numeric_limits<double>::epsilon();
 
   BOOST_CHECK_CLOSE(average_list_of_allowed_entries(std::vector<long>{1, 2}),
@@ -57,7 +57,7 @@ BOOST_AUTO_TEST_CASE(average_list_of_allowed_entries_test) {
 }
 
 BOOST_AUTO_TEST_CASE(factorial_Ni0_divided_by_factorial_Ni0_plus_nu_i_test) {
-  using namespace ReactionEnsemble;
+  using namespace ReactionMethods;
   constexpr double tol = 100 * std::numeric_limits<double>::epsilon();
 
   auto const reaction_ensemble_combinations = [](int N, int nu) {
