+++ src/core/unit_tests/CMakeLists.txt
@@ -46,8 +46,3 @@ unit_test(NAME LocalBox_test SRC LocalBox_test.cpp DEPENDS EspressoCore)
 unit_test(NAME thermostats_test SRC thermostats_test.cpp DEPENDS EspressoCore)
 unit_test(NAME random_test SRC random_test.cpp DEPENDS EspressoUtils Random123)
 unit_test(NAME BondList_test SRC BondList_test.cpp DEPENDS EspressoCore)
-unit_test(NAME reaction_ensemble_utils_test SRC
-          reaction_ensemble_utils_test.cpp DEPENDS EspressoCore)
-unit_test(NAME reaction_ensemble_classes_test SRC
-          reaction_ensemble_classes_test.cpp DEPENDS EspressoCore Boost::mpi
-          MPI::MPI_CXX)
